Update exploit.php
Added options for curl to ignore SSL certs on lines 172 and 173. For some reason the triggerPayload function doesn't ignore certs which is out of line with the other functions and can lead to the script working up to the point it needs to trigger the payload and generate the reverse shell.
Also added extra output on line 184 with details of URL to browse to just in case curl fails for any other reason.